python-projects/12-content-moderation/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
jinja2==3.1.3
websockets==12.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
bcrypt==4.1.2

# Queue System
celery[redis]==5.3.4
redis==4.6.0
flower==2.0.1

# LLM Providers
openai==1.12.0
anthropic==0.18.0
requests==2.31.0
aiohttp==3.9.3

# NSFW Detection
nudenet==3.4.2

# Video Processing
ffmpeg-python==0.2.0
opencv-python==4.9.0.80
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
locust==2.20.0
